constants moved into env
