+++ packages/frontend/.env
@@ -1,3 +1,10 @@
 NEXT_PUBLIC_API_URL=https://testnet.platform-explorer.pshenmic.dev
 NEXT_PUBLIC_BASE_URL=https://testnet.platform-explorer.com
+
+NEXT_PUBLIC_MAINNET_BASE_URL=https://platform-explorer.com
+NEXT_PUBLIC_TESTNET_BASE_URL=https://testnet.platform-explorer.com
+
+NEXT_PUBLIC_MAINNET_INSIGHT_URL=https://insight.dash.org/insight
+NEXT_PUBLIC_TESTNET_INSIGHT_URL=http://insight.testnet.networks.dash.org/insight
+
 NEXT_PUBLIC_VOTING_DATA_CONTRACT_ID=GWRSAVFMjXx8HpQFaNJMqBV7MBgMK4br5UESsB4S31Ec
