+++ packages/frontend/src/constants/networks.js
@@ -8,31 +8,19 @@ export const NETWORK_OPTIONS = {
     name: NETWORKS_ENUM.MAINNET,
     subname: '',
     disabled: false,
-    explorerBaseUrl: 'https://platform-explorer.com',
-    l1explorerBaseUrl: 'https://insight.dash.org/insight'
+    explorerBaseUrl: process.env.NEXT_PUBLIC_MAINNET_BASE_URL,
+    l1explorerBaseUrl: process.env.NEXT_PUBLIC_MAINNET_INSIGHT_URL
   },
   [NETWORKS_ENUM.TESTNET]: {
     name: NETWORKS_ENUM.TESTNET,
     subname: '',
     disabled: false,
-    explorerBaseUrl: 'https://testnet.platform-explorer.com',
-    l1explorerBaseUrl: 'http://insight.testnet.networks.dash.org/insight'
+    explorerBaseUrl: process.env.NEXT_PUBLIC_TESTNET_BASE_URL,
+    l1explorerBaseUrl: process.env.NEXT_PUBLIC_TESTNET_INSIGHT_URL
   }
 }
 
 export const networks = [
-  {
-    name: NETWORKS_ENUM.MAINNET,
-    subname: '',
-    disabled: false,
-    explorerBaseUrl: 'https://platform-explorer.com',
-    l1explorerBaseUrl: 'https://insight.dash.org/insight'
-  },
-  {
-    name: NETWORKS_ENUM.TESTNET,
-    subname: '',
-    disabled: false,
-    explorerBaseUrl: 'https://testnet.platform-explorer.com',
-    l1explorerBaseUrl: 'http://insight.testnet.networks.dash.org/insight'
-  }
+  NETWORK_OPTIONS[NETWORKS_ENUM.MAINNET],
+  NETWORK_OPTIONS[NETWORKS_ENUM.TESTNET]
 ]
